Trait grpc_attestation::proto::streaming_session_server::StreamingSession

pub trait StreamingSession: Send + Sync + 'static {
    type StreamStream: Stream<Item = Result<StreamingResponse, Status>> + Send + 'static;
    fn stream<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Streaming<StreamingRequest>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with StreamingSessionServer.

Associated Types

type StreamStream: Stream<Item = Result<StreamingResponse, Status>> + Send + 'static

Server streaming response type for the Stream method.

Required methods

fn stream<'life0, 'async_trait>(
&'life0 self,
request: Request<Streaming<StreamingRequest>>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Creates a message stream for session key negotiation and encrypted payload exchange.

The created message stream looks as follows:

Client->Server: StreamingRequest with a serialized ClientHello message.

Server->Client: StreamingResponse with a serialized ServerIdentity message.

Client->Server: StreamingRequest with a serialized ClientIdentity message.

Client->Server: StreamingRequest with a serialized EncryptedData message.

Server->Client: StreamingResponse with a serialized EncryptedData message.

Messages are represented as serialized messages defined in the remote_attestation::message.rs and com.google.oak.remote_attestation.Message.